+++ app/policies/dorsale/customer_vault/event_policy_helper.rb
@@ -1,0 +1,4 @@
+module Dorsale::CustomerVault::EventPolicyHelper
+  # Only used for scopes
+  POLICY_METHODS = []
+end