+++ spec/dummy/app/policies/dorsale/customer_vault/event_policy.rb
@@ -1,0 +1,4 @@
+class Dorsale::CustomerVault::EventPolicy < Dorsale::ApplicationPolicy
+  prepend Dorsale::CustomerVault::EventPolicyHelper
+  define_dummy_policies!
+end